+++ radical_translations/utils/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 
-from radical_translations.utils.views import PlaceDetailView
+from radical_translations.utils.views import PlaceRetrieveView
 
 urlpatterns = [
-    path("place/<int:pk>/", PlaceDetailView.as_view(), name="place-detail"),
+    path("api/places/<int:pk>/", PlaceRetrieveView.as_view(), name="api-place-detail"),
 ]