+++ radical_translations/utils/views.py
@@ -1,7 +1,9 @@
-from django.views.generic.detail import DetailView
+from rest_framework.generics import RetrieveAPIView
 
 from geonames_place.models import Place
+from geonames_place.serializers import PlaceSerializer
 
 
-class PlaceDetailView(DetailView):
-    model = Place
+class PlaceRetrieveView(RetrieveAPIView):
+    queryset = Place.objects.all()
+    serializer_class = PlaceSerializer